+++ hardware/src/mempool_group.sv
@@ -3,6 +3,8 @@
 // SPDX-License-Identifier: SHL-0.51
 
 `include "mempool/mempool.svh"
+`include "reqrsp_interface/typedef.svh"
+`include "common_cells/registers.svh"
 
 module mempool_group
   import mempool_pkg::*;
@@ -41,6 +43,12 @@ module mempool_group
   input  logic                            [NumCoresPerGroup-1:0]                 wake_up_i,
   // RO-Cache configuration
   input  `STRUCT_PORT(ro_cache_ctrl_t)                                           ro_cache_ctrl_i,
+  // DMA request
+  input  `STRUCT_PORT(dma_req_t)                                                 dma_req_i,
+  input  logic                                                                   dma_req_valid_i,
+  output logic                                                                   dma_req_ready_o,
+  // DMA status
+  output `STRUCT_PORT(dma_meta_t)                                                dma_meta_o,
    // AXI Interface
   output `STRUCT_VECT(axi_tile_req_t,     [NumAXIMastersPerGroup-1:0])           axi_mst_req_o,
   input  `STRUCT_VECT(axi_tile_resp_t,    [NumAXIMastersPerGroup-1:0])           axi_mst_resp_i
@@ -83,6 +91,13 @@ module mempool_group
   tcdm_slave_resp_t  [NumGroups-1:0][NumTilesPerGroup-1:0] tcdm_slave_resp;
   logic              [NumGroups-1:0][NumTilesPerGroup-1:0] tcdm_slave_resp_valid;
   logic              [NumGroups-1:0][NumTilesPerGroup-1:0] tcdm_slave_resp_ready;
+  // DMA interfaces
+  tcdm_dma_req_t  [NumTilesPerGroup-1:0] tcdm_dma_req;
+  logic           [NumTilesPerGroup-1:0] tcdm_dma_req_valid;
+  logic           [NumTilesPerGroup-1:0] tcdm_dma_req_ready;
+  tcdm_dma_resp_t [NumTilesPerGroup-1:0] tcdm_dma_resp;
+  logic           [NumTilesPerGroup-1:0] tcdm_dma_resp_valid;
+  logic           [NumTilesPerGroup-1:0] tcdm_dma_resp_ready;
 
   // Connect the IOs to the tiles' signals
   assign tcdm_master_resp[NumGroups-1:1]         = tcdm_master_resp_i[NumGroups-1:1];
@@ -95,6 +110,8 @@ module mempool_group
   // AXI interfaces
   axi_tile_req_t  [NumTilesPerGroup-1:0] axi_tile_req;
   axi_tile_resp_t [NumTilesPerGroup-1:0] axi_tile_resp;
+  axi_tile_req_t  [NumDmasPerGroup-1:0]  axi_dma_req;
+  axi_tile_resp_t [NumDmasPerGroup-1:0]  axi_dma_resp;
 
   for (genvar t = 0; unsigned'(t) < NumTilesPerGroup; t++) begin: gen_tiles
     tile_id_t id;
@@ -137,6 +154,13 @@ module mempool_group
       .tcdm_slave_resp_o       (tran_tcdm_slave_resp                           ),
       .tcdm_slave_resp_valid_o (tran_tcdm_slave_resp_valid                     ),
       .tcdm_slave_resp_ready_i (tran_tcdm_slave_resp_ready                     ),
+      // TCDM DMA interfaces
+      .tcdm_dma_req_i          (tcdm_dma_req[t]                                ),
+      .tcdm_dma_req_valid_i    (tcdm_dma_req_valid[t]                          ),
+      .tcdm_dma_req_ready_o    (tcdm_dma_req_ready[t]                          ),
+      .tcdm_dma_resp_o         (tcdm_dma_resp[t]                               ),
+      .tcdm_dma_resp_valid_o   (tcdm_dma_resp_valid[t]                         ),
+      .tcdm_dma_resp_ready_i   (tcdm_dma_resp_ready[t]                         ),
       // AXI interface
       .axi_mst_req_o           (axi_tile_req[t]                                ),
       .axi_mst_resp_i          (axi_tile_resp[t]                               ),
@@ -346,31 +370,31 @@ module mempool_group
   axi_tile_resp_t  [NumAXIMastersPerGroup-1:0] axi_mst_resp;
 
   axi_hier_interco #(
-    .NumSlvPorts    (NumTilesPerGroup     ),
-    .NumMstPorts    (NumAXIMastersPerGroup),
-    .Radix          (AxiHierRadix         ),
-    .EnableCache    (32'hFFFFFFFF         ),
-    .CacheLineWidth (ROCacheLineWidth     ),
-    .CacheSizeByte  (ROCacheSizeByte      ),
-    .CacheSets      (ROCacheSets          ),
-    .AddrWidth      (AddrWidth            ),
-    .DataWidth      (AxiDataWidth         ),
-    .SlvIdWidth     (AxiTileIdWidth       ),
-    .MstIdWidth     (AxiTileIdWidth       ),
-    .UserWidth      (1                    ),
-    .slv_req_t      (axi_tile_req_t       ),
-    .slv_resp_t     (axi_tile_resp_t      ),
-    .mst_req_t      (axi_tile_req_t       ),
-    .mst_resp_t     (axi_tile_resp_t      )
+    .NumSlvPorts    (NumTilesPerGroup+NumDmasPerGroup),
+    .NumMstPorts    (NumAXIMastersPerGroup           ),
+    .Radix          (AxiHierRadix                    ),
+    .EnableCache    (32'hFFFFFFFF                    ),
+    .CacheLineWidth (ROCacheLineWidth                ),
+    .CacheSizeByte  (ROCacheSizeByte                 ),
+    .CacheSets      (ROCacheSets                     ),
+    .AddrWidth      (AddrWidth                       ),
+    .DataWidth      (AxiDataWidth                    ),
+    .SlvIdWidth     (AxiTileIdWidth                  ),
+    .MstIdWidth     (AxiTileIdWidth                  ),
+    .UserWidth      (1                               ),
+    .slv_req_t      (axi_tile_req_t                  ),
+    .slv_resp_t     (axi_tile_resp_t                 ),
+    .mst_req_t      (axi_tile_req_t                  ),
+    .mst_resp_t     (axi_tile_resp_t                 )
   ) i_axi_interco (
-    .clk_i           (clk_i               ),
-    .rst_ni          (rst_ni              ),
-    .test_i          (1'b0                ),
-    .ro_cache_ctrl_i (ro_cache_ctrl_i     ),
-    .slv_req_i       (axi_tile_req        ),
-    .slv_resp_o      (axi_tile_resp       ),
-    .mst_req_o       (axi_mst_req         ),
-    .mst_resp_i      (axi_mst_resp        )
+    .clk_i           (clk_i                       ),
+    .rst_ni          (rst_ni                      ),
+    .test_i          (1'b0                        ),
+    .ro_cache_ctrl_i (ro_cache_ctrl_i             ),
+    .slv_req_i       ({axi_dma_req,axi_tile_req}  ),
+    .slv_resp_o      ({axi_dma_resp,axi_tile_resp}),
+    .mst_req_o       (axi_mst_req                 ),
+    .mst_resp_i      (axi_mst_resp                )
   );
 
   for (genvar m = 0; m < NumAXIMastersPerGroup; m++) begin: gen_axi_group_cuts
@@ -391,5 +415,214 @@ module mempool_group
       .mst_resp_i(axi_mst_resp_i[m])
     );
   end: gen_axi_group_cuts
+
+  /*********
+   *  DMA  *
+   *********/
+  dma_req_t  dma_req_cut;
+  logic      dma_req_cut_valid;
+  logic      dma_req_cut_ready;
+  dma_meta_t dma_meta_cut;
+
+  spill_register #(
+    .T(dma_req_t)
+  ) i_dma_req_register (
+    .clk_i  (clk_i            ),
+    .rst_ni (rst_ni           ),
+    .data_i (dma_req_i        ),
+    .valid_i(dma_req_valid_i  ),
+    .ready_o(dma_req_ready_o  ),
+    .data_o (dma_req_cut      ),
+    .valid_o(dma_req_cut_valid),
+    .ready_i(dma_req_cut_ready)
+  );
+
+  `FF(dma_meta_o, dma_meta_cut, '0, clk_i, rst_ni);
+
+  dma_req_t  [NumDmasPerGroup-1:0] dma_req;
+  logic      [NumDmasPerGroup-1:0] dma_req_valid;
+  logic      [NumDmasPerGroup-1:0] dma_req_ready;
+  dma_meta_t [NumDmasPerGroup-1:0] dma_meta;
+
+  idma_distributed_midend #(
+    .NoMstPorts     (NumDmasPerGroup                   ),
+    .DmaRegionWidth (NumBanksPerGroup*4/NumDmasPerGroup),
+    .burst_req_t    (dma_req_t                         ),
+    .meta_t         (dma_meta_t                        )
+  ) i_idma_distributed_midend (
+    .clk_i       (clk_i            ),
+    .rst_ni      (rst_ni           ),
+    .burst_req_i (dma_req_cut      ),
+    .valid_i     (dma_req_cut_valid),
+    .ready_o     (dma_req_cut_ready),
+    .meta_o      (dma_meta_cut     ),
+    .burst_req_o (dma_req          ),
+    .valid_o     (dma_req_valid    ),
+    .ready_i     (dma_req_ready    ),
+    .meta_i      (dma_meta         )
+  );
+
+  // xbar
+  localparam int unsigned NumRules = 1;
+  typedef struct packed {
+    int unsigned idx;
+    logic [AddrWidth-1:0] start_addr;
+    logic [AddrWidth-1:0] end_addr;
+  } xbar_rule_t;
+  xbar_rule_t [NumRules-1:0] addr_map;
+  assign addr_map = '{
+    '{ // TCDM
+      start_addr: TCDMBaseAddr,
+      end_addr:   TCDMBaseAddr + TCDMSize,
+      idx:        1
+    }
+  };
+
+  `REQRSP_TYPEDEF_ALL(reqrsp, addr_t, axi_data_t, axi_strb_t)
+
+
+  for (genvar d = 0; unsigned'(d) < NumDmasPerGroup; d++) begin: gen_dmas
+    localparam int unsigned a = NumTilesPerGroup + d;
+
+    axi_tile_req_t  axi_dma_premux_req;
+    axi_tile_resp_t axi_dma_premux_resp;
+    axi_tile_req_t  tcdm_req;
+    axi_tile_resp_t tcdm_resp;
+
+    logic backend_idle;
+    logic trans_complete;
+
+    axi_dma_backend #(
+      .DataWidth       (AxiDataWidth   ),
+      .AddrWidth       (AddrWidth      ),
+      .IdWidth         (AxiTileIdWidth ),
+      .AxReqFifoDepth  (2              ),
+      .TransFifoDepth  (1              ),
+      .BufferDepth     (3              ),
+      .axi_req_t       (axi_tile_req_t ),
+      .axi_res_t       (axi_tile_resp_t),
+      .burst_req_t     (dma_req_t      ),
+      .DmaIdWidth      (1              ),
+      .DmaTracing      (0              )
+    ) i_axi_dma_backend (
+      .clk_i            (clk_i                     ),
+      .rst_ni           (rst_ni                    ),
+      .dma_id_i         (1'b0                      ),
+      .axi_dma_req_o    (axi_dma_premux_req        ),
+      .axi_dma_res_i    (axi_dma_premux_resp       ),
+      .burst_req_i      (dma_req[d]                ),
+      .valid_i          (dma_req_valid[d]          ),
+      .ready_o          (dma_req_ready[d]          ),
+      .backend_idle_o   (dma_meta[d].backend_idle  ),
+      .trans_complete_o (dma_meta[d].trans_complete)
+    );
+
+    // ------------------------------------------------------
+    // AXI connection to EXT/TCDM
+    // ------------------------------------------------------
+
+    localparam axi_pkg::xbar_cfg_t XbarCfg = '{
+      NoSlvPorts:         1,
+      NoMstPorts:         2,
+      MaxMstTrans:        8,
+      MaxSlvTrans:        8,
+      FallThrough:        1'b0,
+      LatencyMode:        axi_pkg::CUT_ALL_PORTS,
+      AxiIdWidthSlvPorts: AxiTileIdWidth,
+      AxiIdUsedSlvPorts:  AxiTileIdWidth,
+      UniqueIds:          1'b0,
+      AxiAddrWidth:       AddrWidth,
+      AxiDataWidth:       AxiDataWidth,
+      NoAddrRules:        NumRules
+    };
+
+    axi_xbar #(
+      .Cfg          (XbarCfg        ),
+      .slv_aw_chan_t(axi_tile_aw_t  ),
+      .mst_aw_chan_t(axi_tile_aw_t  ),
+      .w_chan_t     (axi_tile_w_t   ),
+      .slv_b_chan_t (axi_tile_b_t   ),
+      .mst_b_chan_t (axi_tile_b_t   ),
+      .slv_ar_chan_t(axi_tile_ar_t  ),
+      .mst_ar_chan_t(axi_tile_ar_t  ),
+      .slv_r_chan_t (axi_tile_r_t   ),
+      .mst_r_chan_t (axi_tile_r_t   ),
+      .slv_req_t    (axi_tile_req_t ),
+      .slv_resp_t   (axi_tile_resp_t),
+      .mst_req_t    (axi_tile_req_t ),
+      .mst_resp_t   (axi_tile_resp_t),
+      .rule_t       (xbar_rule_t    )
+    ) i_dma_axi_xbar (
+      .clk_i                (clk_i                        ),
+      .rst_ni               (rst_ni                       ),
+      .test_i               (1'b0                         ),
+      .slv_ports_req_i      (axi_dma_premux_req           ),
+      .slv_ports_resp_o     (axi_dma_premux_resp          ),
+      .mst_ports_req_o      ({tcdm_req,axi_dma_req[d]}    ),
+      .mst_ports_resp_i     ({tcdm_resp,axi_dma_resp[d]}  ),
+      .addr_map_i           (addr_map                     ),
+      .en_default_mst_port_i('1                           ),
+      .default_mst_port_i   ('0                           )
+    );
+
+    reqrsp_req_t dma_reqrsp_req;
+    reqrsp_rsp_t dma_reqrsp_rsp;
+    reqrsp_req_t [NumTilesPerDma-1:0] dma_tile_req;
+    reqrsp_rsp_t [NumTilesPerDma-1:0] dma_tile_rsp;
+
+    axi_to_reqrsp #(
+      .axi_req_t   (axi_tile_req_t ),
+      .axi_rsp_t   (axi_tile_resp_t),
+      .AddrWidth   (AddrWidth      ),
+      .DataWidth   (AxiDataWidth   ),
+      .IdWidth     (AxiTileIdWidth ),
+      .BufDepth    (2              ),
+      .reqrsp_req_t(reqrsp_req_t   ),
+      .reqrsp_rsp_t(reqrsp_rsp_t   )
+    ) i_axi_to_reqrsp (
+      .clk_i       (clk_i         ),
+      .rst_ni      (rst_ni        ),
+      .busy_o      (/*unused*/    ),
+      .axi_req_i   (tcdm_req      ),
+      .axi_rsp_o   (tcdm_resp     ),
+      .reqrsp_req_o(dma_reqrsp_req),
+      .reqrsp_rsp_i(dma_reqrsp_rsp)
+    );
+
+    reqrsp_demux #(
+      .NrPorts  (NumTilesPerDma),
+      .req_t    (reqrsp_req_t  ),
+      .rsp_t    (reqrsp_rsp_t  ),
+      .RespDepth(2             )
+    ) i_reqrsp_demux (
+       .clk_i       (clk_i                                                                                  ),
+       .rst_ni      (rst_ni                                                                                 ),
+       .slv_select_i(dma_reqrsp_req.q.addr[idx_width(NumBanksPerTile)+ByteOffset+:idx_width(NumTilesPerDma)]),
+       .slv_req_i   (dma_reqrsp_req                                                                         ),
+       .slv_rsp_o   (dma_reqrsp_rsp                                                                         ),
+       .mst_req_o   (dma_tile_req                                                                           ),
+       .mst_rsp_i   (dma_tile_rsp                                                                           )
+    );
+
+    // Assignment to TCDM interconnect
+    // TODO: Reordering might be problematic
+    for (genvar t = 0; unsigned'(t) < NumTilesPerDma; t++) begin: gen_dma_tile_connection
+      assign tcdm_dma_req[d*NumTilesPerDma+t] = '{
+               wdata: dma_tile_req[t].q.data,
+               wen: dma_tile_req[t].q.write,
+               be: dma_tile_req[t].q.strb,
+               tgt_addr: {dma_tile_req[t].q.addr[ByteOffset + idx_width(NumBanksPerTile) + $clog2(NumTilesPerGroup) + $clog2(NumGroups)+:TCDMAddrMemWidth],
+                          dma_tile_req[t].q.addr[ByteOffset+:idx_width(NumBanksPerTile)]}
+             };
+      assign tcdm_dma_req_valid[d*NumTilesPerDma+t]  = dma_tile_req[t].q_valid;
+      assign dma_tile_rsp[t].q_ready = tcdm_dma_req_ready[d*NumTilesPerDma+t];
+      assign dma_tile_rsp[t].p = '{
+               data: tcdm_dma_resp[d*NumTilesPerDma+t].rdata,
+               error: '0
+             };
+      assign dma_tile_rsp[t].p_valid = tcdm_dma_resp_valid[d*NumTilesPerDma+t];
+      assign tcdm_dma_resp_ready[d*NumTilesPerDma+t] =dma_tile_req[t].p_ready;
+    end
+  end
 
 endmodule: mempool_group