+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Add a DMA
+
+### Fixed
+
+### Changed
+- Increase the default AXI width
+
 ## 0.5.0 - 2022-08-03
 
 ### Added